rpl: uint32_t is sufficient for the DAO timer

+++ sys/include/net/gnrc/rpl/structs.h
@@ -223,7 +223,7 @@ struct gnrc_rpl_dodag {
     bool dodag_conf_requested;      /**< flag to send DODAG_CONF options */
     bool prefix_info_requested;     /**< flag to send PREFIX_INFO options */
     msg_t dao_msg;                  /**< msg_t for firing a dao */
-    uint64_t dao_time;              /**< time to schedule the next DAO */
+    uint32_t dao_time;              /**< time to schedule the next DAO */
     xtimer_t dao_timer;             /**< timer to schedule the next DAO */
     msg_t cleanup_msg;              /**< msg_t for firing a cleanup */
     uint32_t cleanup_time;          /**< time to schedula a DODAG cleanup */
